# packages/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "hunter/seeker API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/messaging/conversations": {
      "get": {
        "tags": [
          "messaging"
        ],
        "summary": "The caller's conversations, most recent first",
        "operationId": "messaging_list_conversations",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from a previous page's `next_cursor`. Omit for the first page.",
              "title": "Cursor"
            },
            "description": "Opaque cursor from a previous page's `next_cursor`. Omit for the first page."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page size, 1-100.",
              "default": 20,
              "title": "Limit"
            },
            "description": "Page size, 1-100."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Conversation_"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/messaging/conversations/{conversation_id}/messages": {
      "post": {
        "tags": [
          "messaging"
        ],
        "summary": "Send a message in a conversation",
        "operationId": "messaging_send_message",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Conversation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/feed": {
      "get": {
        "tags": [
          "feed"
        ],
        "summary": "The calling Seeker's Feed",
        "description": "Posts from the Seeker's Connections interleaved with new Matches, newest first.\n\nInterleaving is done here, server-side; clients render one list and never merge two.\nMatches respect the Seeker's match threshold, as on every seeker-facing surface.",
        "operationId": "feed_list_feed",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from a previous page's `next_cursor`. Omit for the first page.",
              "title": "Cursor"
            },
            "description": "Opaque cursor from a previous page's `next_cursor`. Omit for the first page."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page size, 1-100.",
              "default": 20,
              "title": "Limit"
            },
            "description": "Page size, 1-100."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_FeedItem_"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/feed/posts": {
      "post": {
        "tags": [
          "feed"
        ],
        "summary": "Publish a post",
        "description": "Publish a text post to the caller's Connections' Feeds.",
        "operationId": "feed_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Conversation": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "participants": {
            "items": {
              "$ref": "#/components/schemas/UserSummary"
            },
            "type": "array",
            "minItems": 2,
            "title": "Participants"
          },
          "last_message": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Message"
              },
              {
                "type": "null"
              }
            ]
          },
          "unread_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Unread Count",
            "description": "Unread by the caller."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "participants",
          "unread_count",
          "updated_at"
        ],
        "title": "Conversation"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable, machine-readable identifier, e.g. `not_found`."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable explanation. Not stable; do not parse."
          },
          "details": {
            "$ref": "#/components/schemas/JsonValue",
            "description": "Free-form context, or null."
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "description": "The ``error`` object inside the envelope."
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "description": "Body of every error response."
      },
      "FeedItem": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/Post"
          },
          {
            "$ref": "#/components/schemas/MatchedJobItem"
          }
        ],
        "discriminator": {
          "propertyName": "kind",
          "mapping": {
            "matched_job": "#/components/schemas/MatchedJobItem",
            "post": "#/components/schemas/Post"
          }
        }
      },
...
      "JobSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "company_profile_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Profile Id"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "remote": {
            "type": "boolean",
            "title": "Remote",
            "default": false
          },
          "compensation_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compensation Min",
            "description": "Annual, in `currency`."
          },
          "compensation_max": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compensation Max",
            "description": "Annual, in `currency`."
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "description": "ISO 4217 code.",
            "default": "USD"
          },
          "seniority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seniority"
          },
          "posted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Posted At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "company_profile_id",
          "company_name",
          "posted_at"
        ],
        "title": "JobSummary",
        "description": "A Job as referenced from a Match, an Application, or the Feed."
      },
      "JsonValue": {
        "anyOf": [
          {
            "additionalProperties": true,
            "type": "object"
          },
          {
            "items": {},
            "type": "array"
          },
          {
            "type": "string"
          },
          {
            "type": "integer"
          },
          {
            "type": "number"
          },
          {
            "type": "boolean"
          },
          {
            "type": "null"
          }
        ]
      },
...
      "MatchedJob": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score",
            "description": "Normalized 0-1 Match score."
          },
          "ats_pass": {
            "type": "boolean",
            "const": true,
            "title": "Ats Pass",
            "description": "Always `true`: a Match exists only if the Seeker's Profile has already passed the Job's ATS screening. Present so the invariant is visible in the contract.",
            "default": true
          },
          "computed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Computed At",
            "description": "When `score` was last (re)computed."
          },
          "job": {
            "$ref": "#/components/schemas/JobSummary"
          }
        },
        "type": "object",
        "required": [
          "id",
          "score",
          "computed_at",
          "job"
        ],
        "title": "MatchedJob",
        "description": "Seeker-side view of a Match: the Job it recommends."
      },
      "MatchedJobItem": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "matched_job",
            "title": "Kind",
            "default": "matched_job"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Feed item id, distinct from `match.id`."
          },
          "match": {
            "$ref": "#/components/schemas/MatchedJob"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "When the Match entered the Feed."
          }
        },
        "type": "object",
        "required": [
          "id",
          "match",
          "created_at"
        ],
        "title": "MatchedJobItem"
      },
      "Message": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "conversation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Conversation Id"
          },
          "sender": {
            "$ref": "#/components/schemas/UserSummary"
          },
          "body": {
            "type": "string",
            "maxLength": 10000,
            "title": "Body"
          },
          "sent_at": {
            "type": "string",
            "format": "date-time",
            "title": "Sent At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "conversation_id",
          "sender",
          "body",
          "sent_at"
        ],
        "title": "Message"
      },
      "MessageCreate": {
        "properties": {
          "body": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 1,
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "body"
        ],
        "title": "MessageCreate",
        "description": "Body of ``POST /messaging/conversations/{conversation_id}/messages``."
      },
...
      "Page_Conversation_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Conversation"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "Pass as `cursor` to fetch the next page. `null` on the last page."
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "Page[Conversation]"
      },
      "Page_FeedItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FeedItem"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "Pass as `cursor` to fetch the next page. `null` on the last page."
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "Page[FeedItem]"
      },
...
      "Persona": {
        "type": "string",
        "enum": [
          "seeker",
          "hunter"
        ]
      },
      "Post": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "post",
            "title": "Kind",
            "default": "post"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "author": {
            "$ref": "#/components/schemas/UserSummary"
          },
          "body": {
            "type": "string",
            "maxLength": 5000,
            "title": "Body"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "author",
          "body",
          "created_at"
        ],
        "title": "Post"
      },
      "PostCreate": {
        "properties": {
          "body": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "body"
        ],
        "title": "PostCreate",
        "description": "Body of ``POST /feed/posts``. Text only for now; media and mentions come later."
      },
...
      "UserSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "persona": {
            "$ref": "#/components/schemas/Persona"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name",
            "description": "What other users see. Never a legal name."
          }
        },
        "type": "object",
        "required": [
          "id",
          "persona",
          "display_name"
        ],
        "title": "UserSummary",
        "description": "The minimal public view of a user, embedded wherever a user is referenced."
      },
...
    }
  }
}